+++ src/page-objects/AdministrationPages.ts
@@ -51,6 +51,7 @@ import { LayoutCreate } from "./administration/LayoutCreate";
 import { ListingPageLayoutDetail } from "./administration/ListingPageLayoutDetail";
 import { SalesChannelDetail } from "./administration/SalesChannelDetail";
 import { OrderListing } from "./administration/OrderListing";
+import { DataSharingConsentModal } from "./administration/DataSharingConsentModal";
 
 export interface AdministrationPageTypes {
     AdminProductDetail: ProductDetail;
@@ -103,6 +104,7 @@ export interface AdministrationPageTypes {
     AdminListingPageLayoutDetail: ListingPageLayoutDetail;
     AdminLayoutCreate: LayoutCreate;
     AdminOrderListing: OrderListing;
+    AdminDataSharingConsentModal: DataSharingConsentModal;
 }
 
 export const AdminPageObjects = {
@@ -156,6 +158,7 @@ export const AdminPageObjects = {
     ListingPageLayoutDetail,
     LayoutCreate,
     OrderListing,
+    DataSharingConsentModal,
 };
 
 export const test = base.extend<FixtureTypes>({
@@ -342,8 +345,8 @@ export const test = base.extend<FixtureTypes>({
         await use(new LayoutListing(AdminPage));
     },
 
-    AdminListingPageLayoutDetail: async ({ AdminPage }, use) => {
-        await use(new ListingPageLayoutDetail(AdminPage));
+    AdminListingPageLayoutDetail: async ({ AdminPage, InstanceMeta }, use) => {
+        await use(new ListingPageLayoutDetail(AdminPage, InstanceMeta));
     },
 
     AdminLayoutCreate: async ({ AdminPage }, use) => {
@@ -356,5 +359,9 @@ export const test = base.extend<FixtureTypes>({
 
     AdminOrderListing: async ({ AdminPage, InstanceMeta }, use) => {
         await use(new OrderListing(AdminPage, InstanceMeta));
+    },
+
+    AdminDataSharingConsentModal: async ({ AdminPage }, use) => {
+        await use(new DataSharingConsentModal(AdminPage));
     },
 });