+++ src/page-objects/administration/Dashboard.ts
@@ -12,6 +12,8 @@ export class Dashboard implements PageObject {
     public readonly adminMenuContent: Locator;
     public readonly adminMenuMarketing: Locator;
     public readonly adminMenuExtension: Locator;
+    public readonly adminMenuAutomation: Locator;
+    public readonly adminMenuSettings: Locator;
     public readonly adminMenuUserChevron: Locator;
     public readonly adminMenuUserIcon: Locator;
     public readonly adminMenuUserName: Locator;
@@ -46,9 +48,11 @@ export class Dashboard implements PageObject {
         this.adminMenuContent = page.locator(".sw-content");
         this.adminMenuMarketing = page.locator(".sw-marketing");
         this.adminMenuExtension = page.locator(".sw-extension");
+        this.adminMenuAutomation = page.locator(".sw-automation");
         this.adminMenuUserChevron = page.locator(".sw-admin-menu__user-actions-indicator");
         this.adminMenuUserIcon = page.locator(".sw-avatar");
         this.adminMenuUserName = page.locator(".sw-admin-menu__user-name");
+        this.adminMenuSettings = page.locator(".sw-settings");
         this.welcomeHeadline = page.locator("h1.sw-dashboard-index__welcome-title");
         this.welcomeMessage = page.locator(".sw-dashboard-index__welcome-message");
         this.dataSharingConsentBanner = page.locator(".sw-usage-data-consent-banner");